Added A* Search Code and terminal input
Completed all TODOs and added new files for support functions to
CMakeLists.txt. Input setter was created as a way of taking in
coordinate inputs without repeating the same input message each time.
Node comparer was created so that the compare function was not sitting
in the same file as the route planner code, making the code easier to
follow.